Combine `EmptyClassroomChecker` and `EmptyClassroomCounter` as `EmptyClassroomHelper`.
